+++ test/import_err/err/__init__.py
@@ -1,1 +1,1 @@
-raise RuntimeError("I fail on import")
+raise RuntimeError