+++ CHANGELOG.md
@@ -7,6 +7,7 @@
    on function arguments.
  - The "Edit on GitHub" button now says "Edit on GitLab" if it points to
    GitLab, or "Edit Source" if neither platform is used.
+ - Add `pdoc.doc.Module.from_name` to simplify module creation.
  - Do not linkify identifiers that are already manually linked.
  - When importing local file paths, always make sure that the directory is
    at the front of `sys.path`.